+++ docs/pages/documentation/installation/Customization.vue
@@ -15,10 +15,10 @@
         <hr>
 
         <h3 class="subtitle"><span class="list-number">2</span> Set your variables with Sass</h3>
-        <p class="content">
-            <b>Tip:</b> You can see all Bulma's variables <a href="http://bulma.io/documentation/overview/variables/" target="_blank">here</a>.
-            There's also another variable created by Buefy: <code>$speed-slow: 150ms !default</code>.
-        </p>
+        <div class="content">
+            <p><b>Tip:</b> You can see all Bulma's variables <a href="http://bulma.io/documentation/overview/variables/" target="_blank">here</a>.</p>
+            <p>There's also two others variables created by Buefy: <code>$speed-slow: 150ms !default</code> and <code>$speed-slower: 250ms !default</code>.</p>
+        </div>
         <div class="example"><pre v-highlight><code class="sass">{{ sass | pre }}</code></pre></div>
 
         <hr>